export and load blueprints

+++ Yokai_Engine/Source Code/DN_Int.cpp
@@ -30,5 +30,5 @@ DN_Int::~DN_Int()
 
 void DN_Int::SaveInNode()
 {
-	info_as_number = std::stoi(inputs[0].string_box);
+	info_as_number = (float)std::stoi(inputs[0].string_box);
 }
